requirements.txt
fastapi[standard]
sqlalchemy
python-dotenv
pyjwt
pytest
aiosqlite
pytest-asyncio
httpx